Test and fix mistakes

+++ Module_3/Project/AirTransport.cpp
@@ -6,10 +6,10 @@ AirTransport::AirTransport(std::string name, int velocity, float distanceReducti
 {}
 
 
-int AirTransport::getDistanceReductionFactorInPercent() const {
+float AirTransport::getDistanceReductionFactorInPercent() const {
     return _distanceReductionFactorInPercent;
 }
 
 void AirTransport::setDistanceReductionFactorInPercent(float newFactor){
-    if (newFactor >= 0) _distanceReductionFactorInPercent = newFactor;
+    if (newFactor >= 0) _distanceReductionFactorInPercent = static_cast<int>(newFactor);
 }
